chore: Update todo retrieval in ServerTodosPage

+++ src/app/api/todos/route.ts
@@ -1,6 +1,7 @@
 import { NextResponse } from 'next/server'
 import prisma from '@/lib/prisma'
 import { object, string } from 'yup';
+import { getUserSession } from '@/auth/actions/auth-actions';
 
 export async function GET(request: Request) { 
     const {searchParams} = new URL(request.url);
@@ -32,9 +33,19 @@ const postSchema = object({
 
 export async function POST(request: Request) {
 
+    const user = await getUserSession();
+    if(!user) {
+        return NextResponse.json({message: 'Unauthorized'}, {status: 401});
+    }
+
     try {
         const { description} = await postSchema.validate(await request.json());
-        const todo = await prisma.todo.create({data: { description}});
+        const todo = await prisma.todo.create({
+            data: {
+                description,
+                userId: user.id
+            }
+        });
             
     
         if(!description) {
@@ -49,8 +60,14 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+
+    const user = await getUserSession();
+    if(!user) {
+        return NextResponse.json({message: 'Unauthorized'}, {status: 401});
+    }
+
     try {
-        await prisma.todo.deleteMany({ where: { complete: true } });
+        await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } });
         return NextResponse.json({ message: 'Completed todos deleted successfully' }, { status: 200 });
     } catch (error) {
         console.error('Error deleting completed todos:', error);
